% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_rect.r
\name{lambda_rector}
\alias{lambda_rector}
\title{Correcting for 16S rRNA copy-number based on Lambda phage standard.}
\usage{
lambda_rector(
  ps,
  lamba_id = "Lambda",
  out_path = "./",
  negative_cont = NULL,
  negative_filt = TRUE,
  rare_depth = 10000,
  taxa_level = "Kingdom"
)
}
\arguments{
\item{ps}{a raw phyloseq object.}

\item{lamba_id}{the taxa name assinged to lambda phage in your tax_table of ps file}

\item{out_path}{path to output files}

\item{negative_cont}{If you have used negative controls, a vector of negative control sample names.}

\item{negative_filt}{Wheather data should be filtered based on negative control reads}

\item{rare_depth}{Rarefaction depth}

\item{taxa_level}{Taxonomic level to which analysis should be done.}
}
\description{
Correcting for 16S rRNA copy-number based on Lambda phage standard.
}
